+++ models/PyCryptoBot.py
@@ -97,13 +97,13 @@ class PyCryptoBot(BotConfig):
 
     def _isMarketValid(self, market):
         if self.exchange == Exchange.COINBASEPRO or self.exchange == Exchange.KUCOIN:
-            p = re.compile(r"^[1-9A-Z]{2,5}\-[1-9A-Z]{2,5}$")
+            p = re.compile(r"^[0-9A-Z]{1,10}\-[1-9A-Z]{2,5}$")
             return p.match(market)
         elif self.exchange == Exchange.BINANCE:
-            p = re.compile(r"^[A-Z0-9]{6,12}$")
+            p = re.compile(r"^[A-Z0-9]{5,17}$")
             if p.match(market):
                 return True
-            p = re.compile(r"^[1-9A-Z]{2,5}\-[1-9A-Z]{2,5}$")
+            p = re.compile(r"^[0-9A-Z]{1,10}\-[1-9A-Z]{2,5}$")
             if p.match(market):
                 return True
             return False
@@ -205,6 +205,13 @@ class PyCryptoBot(BotConfig):
         except Exception:  # pylint: disable=broad-except
             return None
 
+    def buyLastSellSize(self) -> bool:
+        return self.buylastsellsize
+
+    def marketMultiBuyCheck(self) -> bool:
+        return self.marketmultibuycheck
+
+
     def getBuyNearHighPcnt(self):
         try:
             return float(self.nobuynearhighpcnt)
@@ -220,11 +227,15 @@ class PyCryptoBot(BotConfig):
         date = date.replace("T", " ") if date.find("T") != -1 else date
         # add time in case only a date is passed in
         new_date_str = f"{date} 00:00:00" if len(date) == 10 else date
-
         return datetime.strptime(new_date_str, "%Y-%m-%d %H:%M:%S")
 
     def getHistoricalData(
-        self, market, granularity: Granularity, websocket, iso8601start="", iso8601end=""
+        self,
+        market,
+        granularity: Granularity,
+        websocket,
+        iso8601start="",
+        iso8601end="",
     ):
         if self.exchange == Exchange.BINANCE:
             api = BPublicAPI(api_url=self.getAPIURL())
@@ -238,19 +249,26 @@ class PyCryptoBot(BotConfig):
                     iso8601end,
                 )
             else:
-                return api.getHistoricalData(
-                    market, granularity, websocket
-                )
-        elif self.exchange == Exchange.KUCOIN:  # returns data from coinbase if not specified
+                return api.getHistoricalData(market, granularity, websocket)
+        elif (
+            self.exchange == Exchange.KUCOIN
+        ):  # returns data from coinbase if not specified
             api = KPublicAPI(api_url=self.getAPIURL())
 
             if iso8601start != "" and iso8601end == "":
                 return api.getHistoricalData(
-                    market, granularity, iso8601start
+                    market,
+                    granularity,
+                    None,
+                    iso8601start,
                 )
             elif iso8601start != "" and iso8601end != "":
                 return api.getHistoricalData(
-                    market, granularity, iso8601start, iso8601end
+                    market,
+                    granularity,
+                    None,
+                    iso8601start,
+                    iso8601end,
                 )
             else:
                 return api.getHistoricalData(market, granularity, websocket)
@@ -273,13 +291,10 @@ class PyCryptoBot(BotConfig):
                     iso8601end,
                 )
             else:
-                return api.getHistoricalData(
-                    market, granularity, websocket
-                )
+                return api.getHistoricalData(market, granularity, websocket)
 
     def getSmartSwitchDataFrame(
         self,
-        app,
         df: pd.DataFrame,
         market,
         granularity: Granularity,
@@ -308,8 +323,8 @@ class PyCryptoBot(BotConfig):
             if df_first is None and df_last is None:
                 text_box = TextBox(80, 26)
 
-                if not app.isSimulation() or (
-                    app.isSimulation() and not app.simResultOnly()
+                if not self.isSimulation() or (
+                    self.isSimulation() and not self.simResultOnly()
                 ):
                     text_box.singleLine()
                     if self.smart_switch:
@@ -373,12 +388,14 @@ class PyCryptoBot(BotConfig):
                         timespec="milliseconds"
                     ) == simstart.isoformat(timespec="milliseconds"):
                         if addingExtraCandles == False:
-                            simstart -= timedelta(minutes=(300 * (granularity.to_integer / 60)))
+                            simstart -= timedelta(
+                                minutes=(300 * (granularity.to_integer / 60))
+                            )
                         addingExtraCandles = True
                         self.extraCandlesFound = True
 
-                if not app.isSimulation() or (
-                    app.isSimulation() and not app.simResultOnly()
+                if not self.isSimulation() or (
+                    self.isSimulation() and not self.simResultOnly()
                 ):
                     text_box.doubleLine()
 
@@ -402,12 +419,11 @@ class PyCryptoBot(BotConfig):
 
     def getSmartSwitchHistoricalDataChained(
         self,
-        _app,
+        market,
+        granularity: Granularity,
         start: str = "",
         end: str = "",
     ) -> pd.DataFrame:
-        market = _app.getMarket()
-        granularity = _app.getGranularity()
 
         if self.isSimulation():
             if _app.getSellSmartSwitch() == 1:
@@ -415,13 +431,13 @@ class PyCryptoBot(BotConfig):
                     _app, self.ema1226_5m_cache, market, Granularity.FIVE_MINUTES, start, end
                 )
             self.ema1226_15m_cache = self.getSmartSwitchDataFrame(
-                _app, self.ema1226_15m_cache, market, Granularity.FIFTEEN_MINUTES, start, end
+                self.ema1226_15m_cache, market, Granularity.FIFTEEN_MINUTES, start, end
             )
             self.ema1226_1h_cache = self.getSmartSwitchDataFrame(
-                _app, self.ema1226_1h_cache, market, Granularity.ONE_HOUR, start, end
+                self.ema1226_1h_cache, market, Granularity.ONE_HOUR, start, end
             )
             self.ema1226_6h_cache = self.getSmartSwitchDataFrame(
-                _app, self.ema1226_6h_cache, market, Granularity.SIX_HOURS, start, end
+                self.ema1226_6h_cache, market, Granularity.SIX_HOURS, start, end
             )
             
             if len(self.ema1226_15m_cache) == 0:
@@ -496,21 +512,23 @@ class PyCryptoBot(BotConfig):
                 return self.ema1226_1h_cache
 
     def getHistoricalDataChained(
-        self, market, granularity: Granularity, max_interations: int = 1
+        self, market, granularity: Granularity, max_iterations: int = 1
     ) -> pd.DataFrame:
         df1 = self.getHistoricalData(market, granularity, None)
 
-        if max_interations == 1:
+        if max_iterations == 1:
             return df1
 
         def getPreviousDateRange(df: pd.DataFrame = None) -> tuple:
-            end_date = df["date"].min() - timedelta(seconds=(granularity.to_integer / 60))
+            end_date = df["date"].min() - timedelta(
+                seconds=(granularity.to_integer / 60)
+            )
             new_start = df["date"].min() - timedelta(hours=300)
             return (str(new_start).replace(" ", "T"), str(end_date).replace(" ", "T"))
 
         iterations = 0
         result_df = pd.DataFrame()
-        while iterations < (max_interations - 1):
+        while iterations < (max_iterations - 1):
             start_date, end_date = getPreviousDateRange(df1)
             df2 = self.getHistoricalData(
                 market, granularity, None, start_date, end_date
@@ -538,15 +556,21 @@ class PyCryptoBot(BotConfig):
                 ].copy()
             elif self.exchange == Exchange.COINBASEPRO:
                 api = CBPublicAPI()
-                df_data = api.getHistoricalData(self.market, Granularity.ONE_HOUR, websocket)
+                df_data = api.getHistoricalData(
+                    self.market, Granularity.ONE_HOUR, websocket
+                )
                 self.ema1226_1h_cache = df_data
             elif self.exchange == Exchange.BINANCE:
                 api = BPublicAPI(api_url=self.getAPIURL())
-                df_data = api.getHistoricalData(self.market, Granularity.ONE_HOUR, websocket)
+                df_data = api.getHistoricalData(
+                    self.market, Granularity.ONE_HOUR, websocket
+                )
                 self.ema1226_1h_cache = df_data
             elif self.exchange == Exchange.KUCOIN:
                 api = KPublicAPI(api_url=self.getAPIURL())
-                df_data = api.getHistoricalData(self.market, Granularity.ONE_HOUR, websocket)
+                df_data = api.getHistoricalData(
+                    self.market, Granularity.ONE_HOUR, websocket
+                )
                 self.ema1226_1h_cache = df_data
             else:
                 return False
@@ -574,15 +598,21 @@ class PyCryptoBot(BotConfig):
                 ].copy()
             elif self.exchange == Exchange.COINBASEPRO:
                 api = CBPublicAPI()
-                df_data = api.getHistoricalData(self.market, Granularity.ONE_HOUR, websocket)
+                df_data = api.getHistoricalData(
+                    self.market, Granularity.ONE_HOUR, websocket
+                )
                 self.sma50200_1h_cache = df_data
             elif self.exchange == Exchange.BINANCE:
                 api = BPublicAPI(api_url=self.getAPIURL())
-                df_data = api.getHistoricalData(self.market, Granularity.ONE_HOUR, websocket)
+                df_data = api.getHistoricalData(
+                    self.market, Granularity.ONE_HOUR, websocket
+                )
                 self.sma50200_1h_cache = df_data
             elif self.exchange == Exchange.KUCOIN:
                 api = KPublicAPI(api_url=self.getAPIURL())
-                df_data = api.getHistoricalData(self.market, Granularity.ONE_HOUR, websocket)
+                df_data = api.getHistoricalData(
+                    self.market, Granularity.ONE_HOUR, websocket
+                )
                 self.sma50200_1h_cache = df_data
             else:
                 return False
@@ -606,13 +636,19 @@ class PyCryptoBot(BotConfig):
         try:
             if self.exchange == Exchange.COINBASEPRO:
                 api = CBPublicAPI()
-                df_data = api.getHistoricalData(self.market, Granularity.ONE_DAY, websocket)
+                df_data = api.getHistoricalData(
+                    self.market, Granularity.ONE_DAY, websocket
+                )
             elif self.exchange == Exchange.BINANCE:
                 api = BPublicAPI(api_url=self.getAPIURL())
-                df_data = api.getHistoricalData(self.market, Granularity.ONE_DAY, websocket)
+                df_data = api.getHistoricalData(
+                    self.market, Granularity.ONE_DAY, websocket
+                )
             elif self.exchange == Exchange.KUCOIN:
                 api = KPublicAPI(api_url=self.getAPIURL())
-                df_data = api.getHistoricalData(self.market, Granularity.ONE_DAY, websocket)
+                df_data = api.getHistoricalData(
+                    self.market, Granularity.ONE_DAY, websocket
+                )
             else:
                 return False  # if there is an API issue, default to False to avoid hard sells
 
@@ -637,15 +673,21 @@ class PyCryptoBot(BotConfig):
                 ].copy()
             elif self.exchange == Exchange.COINBASEPRO:
                 api = CBPublicAPI()
-                df_data = api.getHistoricalData(self.market, Granularity.SIX_HOURS, websocket)
+                df_data = api.getHistoricalData(
+                    self.market, Granularity.SIX_HOURS, websocket
+                )
                 self.ema1226_6h_cache = df_data
             elif self.exchange == Exchange.BINANCE:
                 api = BPublicAPI(api_url=self.getAPIURL())
-                df_data = api.getHistoricalData(self.market, Granularity.SIX_HOURS, websocket)
+                df_data = api.getHistoricalData(
+                    self.market, Granularity.SIX_HOURS, websocket
+                )
                 self.ema1226_6h_cache = df_data
             elif self.exchange == Exchange.KUCOIN:
                 api = KPublicAPI(api_url=self.getAPIURL())
-                df_data = api.getHistoricalData(self.market, Granularity.SIX_HOURS, websocket)
+                df_data = api.getHistoricalData(
+                    self.market, Granularity.SIX_HOURS, websocket
+                )
                 self.ema1226_6h_cache = df_data
             else:
                 return False
@@ -669,13 +711,19 @@ class PyCryptoBot(BotConfig):
         try:
             if self.exchange == Exchange.COINBASEPRO:
                 api = CBPublicAPI()
-                df_data = api.getHistoricalData(self.market, Granularity.SIX_HOURS, websocket)
+                df_data = api.getHistoricalData(
+                    self.market, Granularity.SIX_HOURS, websocket
+                )
             elif self.exchange == Exchange.BINANCE:
                 api = BPublicAPI(api_url=self.getAPIURL())
-                df_data = api.getHistoricalData(self.market, Granularity.SIX_HOURS, websocket)
+                df_data = api.getHistoricalData(
+                    self.market, Granularity.SIX_HOURS, websocket
+                )
             elif self.exchange == Exchange.KUCOIN:
                 api = KPublicAPI(api_url=self.getAPIURL())
-                df_data = api.getHistoricalData(self.market, Granularity.SIX_HOURS, websocket)
+                df_data = api.getHistoricalData(
+                    self.market, Granularity.SIX_HOURS, websocket
+                )
             else:
                 return False
 
@@ -1032,7 +1080,7 @@ class PyCryptoBot(BotConfig):
                     self.getAPIPassphrase(),
                     self.getAPIURL(),
                 )
-                return api.marketBuy(market, float(truncate(quote_currency, 2)))
+                return api.marketBuy(market, float(truncate(quote_currency, 8)))
             elif self.exchange == Exchange.KUCOIN:
                 api = KAuthAPI(
                     self.getAPIKey(),
@@ -1040,7 +1088,7 @@ class PyCryptoBot(BotConfig):
                     self.getAPIPassphrase(),
                     self.getAPIURL(),
                 )
-                return api.marketBuy(market, float(truncate(quote_currency, 2)))
+                return api.marketBuy(market, float(truncate(quote_currency, 8)))
             elif self.exchange == Exchange.BINANCE:
                 api = BAuthAPI(
                     self.getAPIKey(),
@@ -1115,8 +1163,8 @@ class PyCryptoBot(BotConfig):
     def startApp(self, app, account, last_action="", banner=True):
         if (
             banner
-            and not app.isSimulation()
-            or (app.isSimulation() and not app.simResultOnly())
+            and not self.isSimulation()
+            or (self.isSimulation() and not self.simResultOnly())
         ):
             self._generate_banner()
 
@@ -1166,7 +1214,9 @@ class PyCryptoBot(BotConfig):
                         endDate -= timedelta(hours=random.randint(0, 8760 * 1))
 
                     startDate = self.getDateFromISO8601Str(str(endDate))
-                    startDate -= timedelta(minutes=(self.getGranularity().to_integer / 60) * 300)
+                    startDate -= timedelta(
+                        minutes=(self.getGranularity().to_integer / 60) * 300
+                    )
 
                 while len(tradingData) < 300 and attempts < 10:
                     if endDate.isoformat() > datetime.now().isoformat():
@@ -1180,7 +1230,6 @@ class PyCryptoBot(BotConfig):
 
                     else:
                         tradingData = self.getSmartSwitchDataFrame(
-                            app,
                             tradingData,
                             self.market,
                             self.getGranularity(),
@@ -1220,30 +1269,35 @@ class PyCryptoBot(BotConfig):
                 tradingData = pd.DataFrame()
 
                 startDate = self.getDateFromISO8601Str(str(datetime.now()))
-                startDate -= timedelta(minutes=(self.getGranularity().to_integer / 60) * 2)
+                startDate -= timedelta(
+                    minutes=(self.getGranularity().to_integer / 60) * 2
+                )
                 endDate = startDate
                 startDate = pd.Series(startDate).dt.round(freq="H")[0]
                 endDate = pd.Series(endDate).dt.round(freq="H")[0]
-                startDate -= timedelta(minutes=(self.getGranularity().to_integer / 60) * 300)
+                startDate -= timedelta(
+                    minutes=(self.getGranularity().to_integer / 60) * 300
+                )
 
                 if endDate.isoformat() > datetime.now().isoformat():
                     endDate = datetime.now()
 
                 if self.smart_switch == 1:
                     tradingData = self.getSmartSwitchHistoricalDataChained(
-                        app,
+                        _app,
                         str(startDate),
                         str(endDate),
                     )
                 else:
                     tradingData = self.getSmartSwitchDataFrame(
-                        app,
+                        _app,
                         tradingData,
                         self.getMarket(),
                         self.getGranularity(),
                         self.getDateFromISO8601Str(str(startDate)).isoformat(),
                         endDate.isoformat(),
                     )
+
                 if self.extraCandlesFound:
                     self.simstartdate = str(pd.Series(startDate).dt.round(freq="H")[0])
                     self.simenddate = str(pd.Series(endDate).dt.round(freq="H")[0])
@@ -1385,7 +1439,8 @@ class PyCryptoBot(BotConfig):
         if not self.disabletelegram:
             text_box.line(
                 "Telegram error msgs",
-                str(not self.disableTelegramErrorMsgs()) + " --disabletelegramerrormsgs",
+                str(not self.disableTelegramErrorMsgs())
+                + " --disabletelegramerrormsgs",
             )
 
         text_box.line("Log", str(not self.disableLog()) + "  --disablelog")
@@ -1399,7 +1454,7 @@ class PyCryptoBot(BotConfig):
         )
         text_box.line(
             "Log Buy and Sell orders in JSON",
-            str(self.logbuysellinjson) + "  --logbuysellinjson"
+            str(self.logbuysellinjson) + "  --logbuysellinjson",
         )
 
         if self.getBuyMaxSize():
@@ -1407,9 +1462,21 @@ class PyCryptoBot(BotConfig):
                 "Max Buy Size", str(self.getBuyMaxSize()) + "  --buymaxsize <size>"
             )
 
+
         if self.getBuyMinSize():
             text_box.line(
                 "Min Buy Size", str(self.getBuyMinSize()) + "  --buyminsize <size>"
+
+        if self.buyLastSellSize():
+            text_box.line(
+                "Buy Last Sell Size",
+                str(self.buyLastSellSize()) + "  --buylastsellsize",
+            )
+
+        if self.marketMultiBuyCheck():
+            text_box.line(
+                "Check for Market Multiple Buys",
+                str(self.marketMultiBuyCheck()) + "  --marketmultibuycheck",
             )
 
         if self.disablebuyema and self.disablebuymacd: